requirements.txt
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachelib==0.13.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gevent==25.9.1
geventhttpclient==2.3.7
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.78.0
h11==0.16.0
har2locust==0.9.3
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.43.2
MarkupSafe==3.0.3
msgpack==1.1.2
msgspec==0.20.0
mysql-connector-python==9.5.0
mysqlclient==2.2.7
opentelemetry-api==1.39.1
opentelemetry-distro==0.60b1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asyncio==0.60b1
opentelemetry-instrumentation-click==0.60b1
opentelemetry-instrumentation-dbapi==0.60b1
opentelemetry-instrumentation-flask==0.60b1
opentelemetry-instrumentation-grpc==0.60b1
opentelemetry-instrumentation-jinja2==0.60b1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-mysql==0.60b1
opentelemetry-instrumentation-mysqlclient==0.60b1
opentelemetry-instrumentation-pymongo==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-sqlite3==0.60b1
opentelemetry-instrumentation-system-metrics==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-urllib==0.60b1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-wsgi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
prometheus_client==0.24.1
prometheus_flask_exporter==0.23.2
protobuf==6.33.5
psutil==7.2.2
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
pyzmq==27.1.0
requests==2.32.5
ruff==0.15.0
simple-websocket==1.1.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
Werkzeug==3.1.5
wrapt==1.17.3
wsproto==1.3.2
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
